Initial version of subtitle parsing (+ fmt)
So I made a fairly simple type to represent subtitles and just pulled in
a lib and converted to that type. I actually initially went with a
different lab (subparse) and rewrote it because it was old and a bit
clunky but my abstracted type stayed which says to me it's good enough
_ship it_

Also, have you heard of our lord and saviour cargo fmt?